requirements.txt
asgiref==3.5.2
Django==4.1.4
django-sass-processor==1.2.2
django-widget-tweaks==1.4.12
libsass==0.22.0
sqlparse==0.4.3
langchain
openai
langchain_openai
xhtml2pdf
django-storages
google-cloud-storage
tiktoken
PyPDF2
python-dotenv
faiss-cpu
chardet
langchain-experimental
gunicorn
whitenoise
pandas
pandasai
matplotlib
tabulate
xlrd
mysqlclient
pymysql
openpyxl
flask-sqlalchemy
seaborn 
networkx
langgraph
whitenoise
channels
daphne
asgiref